refactor: improve type safety by replacing any types
- Replace  with proper  type in UsageChart
- Replace  in CustomTooltip with proper typed destructuring
- Move DashboardSummary and UsageTrendData interfaces to main types.ts
- Make UsageTrendData extend ChartDataPoint for type compatibility
- Remove duplicate interface definitions from dashboardMocks.ts

Ensures proper TypeScript type safety throughout the dashboard components
